+++ src/electronic/electronic.js
@@ -7,7 +7,7 @@ class Electronic {
    * @param {number} port - 电子模块port口
    * @param {number} slot - 电子模块slot口
    */
-  constructor(port, slot) {
+  constructor() {
 
   }
 }